feat(refactor): refactored for greater SOC
